requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
pyTelegramBotAPI==4.14.0
python-decouple==3.8
requests==2.31.0
urllib3==2.1.0
